Compass working on timer base
